// src/models/callback-error.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bet callback API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
export interface CallbackError {
    code: string;
    data?: { [key: string]: any; };
}